Rework target acquisition so it caches objects per frame
Game Objects are now cached per frame to improve performance
